back-end/requirements.txt
# Fichier requirements.txt
python-dotenv == 1.0.1
ultralytics==8.1.34
opencv-python==4.7.0.72
uvicorn=0.30.3
FastAPI== 0.111.1
#pip install -r requirements.txt
#numpy.__version__ || pip list || pip show numpy || pip freeze | findstr numpy